backend/requirements.txt
astroid==2.2.5
cffi==1.13.2
Click==7.0
cryptography==2.8
ecdsa==0.13.3
entrypoints==0.3
flake8==3.7.9
flake8-docstrings==1.5.0
flake8-import-order==0.18.1
Flask==1.0.2
Flask-Cors==3.0.8
Flask-SQLAlchemy==2.4.0
future==0.17.1
isort==4.3.18
itsdangerous==1.1.0
Jinja2==2.10.1
jwt==0.6.1
lazy-object-proxy==1.4.0
MarkupSafe==1.1.1
mccabe==0.6.1
pycodestyle==2.5.0
pycparser==2.19
pycryptodome==3.6.6
pydocstyle==5.0.1
pyflakes==2.1.1
pylint==2.3.1
python-jose-cryptodome==1.3.2
six==1.12.0
snowballstemmer==2.0.0
SQLAlchemy==1.3.3
typed-ast==1.3.5
Werkzeug==0.15.3
wrapt==1.11.1
